refactor(execution): put all create for testing methods under testing config
