+++ crates/blockifier/src/transaction/transactions.rs
@@ -1,7 +1,6 @@
 use std::sync::Arc;
 
 use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
-use starknet_api::calldata;
 use starknet_api::core::{ClassHash, CompiledClassHash, ContractAddress, Nonce};
 use starknet_api::deprecated_contract_class::EntryPointType;
 use starknet_api::transaction::{
@@ -15,7 +14,6 @@ use starknet_api::transaction::{
     TransactionSignature,
     TransactionVersion,
 };
-use starknet_types_core::felt::Felt;
 
 use crate::abi::abi_utils::selector_from_name;
 use crate::context::{BlockContext, TransactionContext};
@@ -506,23 +504,6 @@ impl L1HandlerTransaction {
     pub fn payload_size(&self) -> usize {
         // The calldata includes the "from" field, which is not a part of the payload.
         self.tx.calldata.0.len() - 1
-    }
-
-    pub fn create_for_testing(l1_fee: Fee, contract_address: ContractAddress) -> Self {
-        let calldata = calldata![
-            Felt::from(0x123), // from_address.
-            Felt::from(0x876), // key.
-            Felt::from(0x44)   // value.
-        ];
-        let tx = starknet_api::transaction::L1HandlerTransaction {
-            version: TransactionVersion::ZERO,
-            nonce: Nonce::default(),
-            contract_address,
-            entry_point_selector: selector_from_name("l1_handler_set_value"),
-            calldata,
-        };
-        let tx_hash = TransactionHash::default();
-        Self { tx, tx_hash, paid_fee_on_l1: l1_fee }
     }
 }
