+++ crates/gateway/src/config.rs
@@ -209,6 +209,7 @@ impl Default for ChainInfoConfig {
 }
 
 impl ChainInfoConfig {
+    #[cfg(any(test, feature = "testing"))]
     pub fn create_for_testing() -> Self {
         BlockContext::create_for_testing().chain_info().clone().into()
     }
@@ -286,6 +287,7 @@ impl SerializeConfig for StatefulTransactionValidatorConfig {
 }
 
 impl StatefulTransactionValidatorConfig {
+    #[cfg(any(test, feature = "testing"))]
     pub fn create_for_testing() -> Self {
         StatefulTransactionValidatorConfig {
             max_nonce_for_validation_skip: Default::default(),
